Extracted initial code from concat
Similar code was copied across many projects. Eventually, it was turned into
enough of a library that we centralized it in
[concat](https://github.com/compiling-to-categories/concat), since that was the
top of the dependency tree.

However, it’s really independent of any particular Haskell project, so including
it in one only serves to slow down improvements needed by others.